Clarify dcos-metrics role and routes
This PR is the result of applying #831 to staging instead of develop and
squashing all 12 commits into one:

Use statsd-emitter, not echo, for test app

Statsd-emitter is built into DC/OS and is specifically designed for
testing dcos-metrics.

Update metrics quickstart to use DC/OS CLI

The old article was unaware of the CLI and encouraged users to ssh to
invidual nodes and curl localhost. This was very complex and confusing.

Clarify the intention of the metrics API

Polling the Metrics API recursively is not a good use of the API. We
should avoid encouraging users to do so.

Clarify the routes to masters and agents

Users have been confused by the reference to adminrouter on a specific
node. Making it clear that there is a route to agents via master
(adminrouter) should help.

Bring datadog docs into line

Remove optional jq installation

Backport metrics docs update to 1.9

Backport metrics docs update to 1.10

Remove beta flag from 1.11 docs

Remove beta flag from 1.11 docs

Add CLI as prerequisite, instead of 'DC/OS installed'

Remove confusing reference to 'standalone' in datadog plugin