Update index.md
This is labeled a "Quick Start Guide", but it is not. The installation section should be in the Install page, and the rest of this page is a command reference. I have relabeled the page and moved the install information to the Install page, where applicable. This page should be moved to another directory.

A Quick Start Guide should be a mini-version, without extraneous detail, of the entire doc suite or manual. If you want to create one, let me know and I can help.