Incomplete section
The folder name for this doc is incorrect: change "daigonostictools" to "diagnostictools" or "diagnostic-tools".
Line 21 contains this line: "For example, here’s a sandbox view of a prometheus-0-node task from the above list:" But there is no view. Did you mean to include a screenshot of some kind?
Line 63: "if two tasks match the task filter then we see the following error:" There is no error message.
Line 65: "Therefore we need to be more specific:" Again, no text or image follows this.