Updated Mount Disk Resources instructions to use cache system.
We now recommend to move the previous Mesos resources file to a cache
instead of removing it. Following the DC/OS update for 1.10/1.11/1.12,
this cache will automatically be used when creating the new resources
file and thus fix the issue described in DCOS_OSS-3921.

An additional line explaining the goal of this operation has been
added to the 1.12 docs but not backported. This is because the fix
in DCOS_OSS-3921 has been backported to 1.10 and 1.11 but does not
concern all 1.10 and 1.11 versions (but it will be in DC/OS 1.12.0).