Here are my suggestions
Some of the one-line summaries and JIRA tickets were hard to parse, so pleas pardon my ignorance. 
I tried to make the section align a little better with what we are trying to converge upon in DC/OS release notes (for example, "notable changes" is subjective where "new features" is not and it doesn't seem quite right to include the same feature in more than one place). I tried to make all of the entries read as sentences. You can back out my changes, if needed.