improve kommander docs
a couple of systematic changes happened here.

* removed every occurrence of `except goes here`. if the excerpt-frontmatter is
empty metalsmith will not render anything there, so that change seemed ok to me
until we come up with reasonable excerpts.

* subpages have been removed in favor of properly indented "main documents".
there's been a lot of duplication before on e.g. `/administrating/*` where texts
have been under a headline (e.g. "access control") only to find a site with that
name containing the same content. as there's a sidebar on the right containing
all the `h1` and `h2`s it seemed reasonable to deduplicate those things and go
with the convention that the docs-system seems to prefer as it does not provide
a great UX when adding an "empty" root-page (like aforementioned "access
control") to structure the sidebar on the left.

* lists explaining glossary terms now should have the term itself in **bold** -
the only reason being that we did that for some things, i liked the visual
appeal, and i also like consistency.

* texts have been structured for markdown.

* code has been put in code-blocks.

* links have been properly linked.