Fix feedback button query
This is a workaround fix for the feedback button not properly setting the Doc team when linking out to Jira. 

There's a bug in the redirect server that duplicates and appends the entire query to the URL string so JIRA bugs out and doesn't know to set `team=44?pid=10415&yadayada` (@jaxesn has more info) but this change should solve the immediate problem until we sort out the redirect issue.