chore: fix the revision-plugin
to speed up builds, we now make use of the revision-plugin and
assume that all files that need processing are changed after the
markdown-step. the revision-plugin will the generate a hash out of the
current state of that file and not apply further transformers in case
the hash is the same as in the previous run.