chore: improve speed when using RPP
we formerly used the RENDERING_PATH_PATTERN or RPP env var to speed up
local builds.

we did that by not applying layouts to pages that are outside of the
given RPP.

this changes the approach to hiding all .md-files outside the RPP from
Metalsmith altogether so that each processor only has to deal with the
limited amount of files (and not only the layout-processor).

This speeds up the "boot time" to ~10 seconds where it formerly
took ~50 (on my machine).
The same goes for recompilations, which are faster by an equal
proportion.

A nice side effect of this implementation is that the main menu is now
present and functional despite using a RPP.