chore: get rid of the asset branch
we can now performantly handle everything in one go. to have
live-reloading when changing assets, we want to tell `watch` a
`livereload: true`. unfortunately that would open another
live-reload-server. as an optimization we now have a single watch and
thus only need a single live-reload server. everything seems to be as
performant as before and we can remove `metalsmith-branch` later on.