don't force horizontal scrollbar on non-touch-devices
there seems to be no need to set the width of a block element to the
full viewport width. even more so as the drawback is that every device
that's not in "touch mode" (try not using your macs trackpad but an
external mouse) will have a horizontal scrollbar because 100vw does not
account for the vertical scrollbar.