fix docker builds
the layouts are written in a way that requires the latest version of any
given doc to be present during build time. the DO_NOT_BUILD sections
exclude dcos 2.0 since 771142bbdce3c03bbf07cdf5f10b84f01d4c7c81.
unfortunately that version is needed to properly build the docs for the
`dispatch` and `kommander` branches.