fix robots-txt
While having a look at an issue that led to the docs not being able to
be built when one specifies a git_branch, i saw that line and wondered
whether it would be able to work.

As per google (https://developers.google.com/search/reference/robots_txt)
it shouldn't, as we need a line separator inbetween statements. A space
is not enough here.

This may mean that all of our stagings will be indexed by google. Here's
a current instance that won't behave as intended: http://docs-staging.d2iq.com/robots.txt