chore: don't ban crawlers from production for now
unfortunately the mechanism that has been "fixed" recently seemed to
never have worked in the first place. during the build-time for
production `GIT_BRANCH` does not seem to equal "master" and thus we now
have a valid robots.txt on docs.d2iq.com
(https://docs.d2iq.com/robots.txt) which effectively tells google now to
not index the docs anymore.

this is a hotfix as to not lose scoring on google. we should follow up
with a fix that properly adds a robot.txt on all non-prod-instances but
NOT on production.