remove some code that has no effect
we're removing `headerSphereItems` (and related stuff) from all the
layouts, as the only thing these variables were supposed to do is to
conditionally render the ksphere/mesosphere-menus in the upper right.
apparently it was never necessary to do that in the first place as the
statements are only present in the places that shall render those menus.
thus it never was a problem that `headerSphereItems` is always truthy
(even when no according items were found).