declutter layouts
in order to be more flexible with new products and reduce maintenance
cost, we strive for more consistent and simple layouts.

this is a first step to get rid of the `isProduct`-variables that make
it really easy to write spaghetti code (see `header.pug` for an example).

in a next step we might be able to use a single header-layout instead of
needing to duplicate a lot of things for each new product.