update the bubbles-svg and add view all
it's actually pretty easy to manipulate the svg in order to have
properly scaled bubbles with the dimensions we have here. the bubbles
were formerly created to fit an almost quadratic box. we have a really
wide box and thus they were scaled down a lot.

also adding a view all button as requested in the comments.