add spell checking
Add `markdown-spellcheck` as that brought the best results quickly.

Also introduce npm-scripts for convenience:

* `npm run spellcheck` - generates a report showing all offences
* `npm run fixtypos` - invoke an interactive mode to fix typos

Have a look at the newly introduced README section on how we plan to get
rid of errors that are already present.

Now for the bad part:

As we're using a lot of technical terms it's likely that PR authors will
have their PRs marked red by our CI. They can click on the check to see
what exactly the checker is complaining about and can use the tasks
above to correct the offences. the interactive mode offers adding
exceptions to the `.spelling` file.
This will obviously cause some friction and authors will need to know
about how to best get their PR green. So let's make sure that the README
section is comprehensive enough to be helpful for authors before merging
this.