speed up live editing in docker
...and on bare metal.

1. the consistency-flag is mostly about optimizing performance on osx. we
trade off file consistency between the docker host and the container for
speed. we actually live dangerously here and say that the container is
authoritative (via "delegated"). this allows the container to perform
everything needed for a live reload way faster though and has never led
to a loss of information in past projects for me.

2. introducing yet another watch-rule for pages. the way the
watch-plugin is implemented limits file access in subsequent
metalsmith-runs to only those files that are on the right side of those
rules. because we need to access the `layouts`-folder whenever we change
a page, we formerly decided to just pump all markdown files
(`**/*.{md,tmpl}`) through the pipeline again. we now try to be more
clever and introduce a rule that always pumps the layout-files through
the pipeline whenever some page changed (unfortunately the plugin does
not seem to have a means to express that more intention-revealing).

  the main benefit here is that the live-reload server does not send ~6k
  events to the browser before that reloads anymore. a process which
  seems to be much slower in a docker-environment.