enable incremental builds of swagger and ngindox
this allows to quickly update those includes when adding a new one.
we formerly needed to delete all artifacts and recompile them, which
would take several minutes (like 5 on my machine).

we now evade the complexity of those scripts altogether and solve the
compilation in the Makefile. as an additional benefit we can now make
use of modern computers and compute all artifacts in parallel, which
IMMENSELY speeds up this process for me. that may lead to faster
CI-cycles and deployments as well.