trim down algolia-plugin
we have a DO_NOT_INDEX-config that's not in use anymore.
fortunately we only index things that are not present in DO_NOT_BUILD.

thus we now could remove all the functionality to figure out what what
NOT to index.

we also moved the algolia-env-vars to their respective place in the
middleware to clutter the index-file less.

and then we also hardcode the algolia projectId and public key. no
worries, we still can grep for all places. the ALGOLIA_INDEX was not in
use.