refactor hierarchy-plugin and sidebar
the sidebar is the single largest time consumer in our build process.
in order to remedy that at some point in time, this commit tries to make
the plugin and the sidebar a bit more approachable by refactoring to a
more intention revealing style as well as documenting interesting bits
of the current implementation.

this slightly speeds up the build but is mainly an enabler for
subsequent optimizations.