remove obsolete stuff
most things in here relate to (the currently far from functional)
PDF-generation.

we also remove `ci/test` as that's currently not in use, can be deduced
by `ci/site` and will be obsolete with moving the site to S3 anyways.

there were a couple helper-scripts that could have some use. as they're
not docs-specific and we don't use them in our tooling i removed those
anyway:

* scripts/docker-purge.sh
* scripts/reduce-pages.sh
* scripts/update-links.sh