abstract language and isService
we remove most of the differentiation between languages and whether
something is a service or not.

we'll likely change the way our getPathParts-helper works to make it yet
easier to work with services. but for now everything seems to work fine
and we got rid of a whole lot of code.