fix print button
due to some recent changes that add a target=_blank to every
non-internal link, the print button broke.

you'd click it and it would navigate away to an empty page and then try
to print that empty page.

we now replaced the anchor with a button, as buttons don't suffer from
that target-changing we do.