don't build dedicated swagger-docs
why? because they were only ever rendered to be included in PDFs.
so currently we're waiting like 10 minutes on every deployment (and on
every dev's machine that's setup to render docs) for no good reason.

if we want to build PDFs in the futue, we'd most likely do so by using
puppeteer or wkhtml2pdf, so even then this output is of no use.