steps taken:
* search for `docs.d2iq.com.*/ksphere/` and replace within the results.
* search for `("/ksphere/` and replace within the results. (usual markdown links)
* search for `: /ksphere/` and replace within the results. (reference-style links)
* search for `"ksphere/` and replace within the results. (some internal code)
* search for `"/ksphere/` and replace within the results. (some internal code)
* search for `'ksphere/` and replace within the results. (some internal code)
* search for `'/ksphere/` and replace within the results. (some internal code)
* manually adapt the redirects
* manually sieve through the remaining results and change where
appropriate.

there's a lot of space for errors here. i don't have any good advice on
how to review this. either we just try to go with it and you click
around a bit locally or you try to resemble something like the above
proceduce to see if you come up with the same result :man_shrugging: