add a notice about the dcos-ui needed to view columns
the touched columns are not always present on the according clusters and
now have an asterisk that links to a small info-box below the table
explaining just that.

in order to render markdown within html, the surrounding html needs to
be an inline element. we thus changed the message to `span`. and to get
rid of the label that was put in front automatically we refactored the
`message` shortcode along (there's no usage of `fill` anywhere - thus we
got rid of it).