Example Project Object
Adding an example Kubernetes YAML Project Object since the page references the object but does not document or give any example. A full object specification should be added later or in a completely separate documentation section describing how to declaratively create all of these Kommander objects but something is better than nothing, especially if we're going to mention it in the documentation.