one search page to rule them all
* rename `getPathParts` to `getPathInfo` and let it return a
  `searchScope` (something like "Konvoy 1.6").
* add hints regarding this searchScope on every landingpage, so it can
  search within the newest version of our product (i consider this an
  enhancement over the previous behaviour that searched all versions -
  let me know if that sounds wrong).

some reasoning for this change:

due to an intermediate state of this work it seemed worthwhile to unify
all search pages. the idea is that all directories within a product
folder (as `/pages/dkp/konvoy`) are relevant search-scopes.

unfortunately we currently have like 15 irrelevant "scopes" as e.g.
`Konvoy - Search` (which stems from `/pages/dkp/konvoy/search`)is not as
good as a scope as `Konvoy - 1.6`. Therefore we now remove those
directories in favor of a single search site. once we removed all
other irrelevant scopes we can do stuff like automatically include new
product versions when dirs are added, which would help tremenduously
with maintaining this repo.