decrease build size
a lot of the markup on our pages belong to the sidebar, as we always
render all entries and each of those has really long classnames attached
to it.

`du -sm` tells us `1482MB` before and `903MB` after this change.
if we take into account that images are `749MB` (and thus are our next
target), we get a html-size reduction of about 80%. the size of e.g.
`mesosphere/dcos/2.2/release-notes/2.2.1/` went from 160KB to 15KB.

this change is twofold:

1. don't render nested children that would be invisible
2. don't use BEM for sidebar-CSS as the long class names are the culprit
   here.