less redirecting
when having a look at the dev-console on production we notice that no
link in the sidebar has a trailing slash, which lets S3 do a redirect to
the slashed version.

on my connection that makes about half of the total loading time for
small to medium sized pages (mostly all those without images).
let's cut loading times in half!