only load swagger.js when we need it
we currently load and interpret this 1.3MB monster on every page. that's
especially noticeable when browsing through pages via the sidebar or on
the very first load of docs (e.g. google ;)). the code is interpreted
synchronuously and thus blocks rendering.

let's only load the swagger JS when a swagger-component is on the
current page.

this saves 400k (about 35%) of bandwidth on most pages. it also makes a
very noticeable difference in when the page seems ready to interact
with.