remove addThis
addThis is a service that makes for a great share-button.
unfortunately they add 9 script files in total that weight in at ~1MB
and they add a click handler that catches clicks on internal and
external links which makes for a noticeable delay in changing pages.

as we guess that this button is not used very often, we remove it for
now. if we ever want to have share functionality, that's pretty easy
(and also stable by now) to add for the major platforms even without
using a service.

closes D2IQ-73323