shrink favicon
let's serve favicon's per environment. that way not all devices need to
download those 150k but might be fine with downloading 1 or 15k - as we
have a no-cache policy on the favicon which is the only part that's
re-downloaded on a recurring visit, this directly impacts bandwidth. we
might shrink the bandwidth on a second visit (with a regular
desktop-browser) to <2kb altogether. :tada: