faster rebuilds
we formerly could not rerender a single file as the hierarchy-plugin
only knows about pages that are in the current "run" (or compilation
step).

the watch-plugin will only throw those files into that run, that are
selected via `paths[WATCH_EXPR].value`, e.g. `"**/*.md"`. That means
that we always! threw all! markdown-files into the run to rebuild any
single file.

We now change the hierarchy plugin to cache inbetween runs, so we can
just recompile a single file (which is denoted by the `true`-value).

We introduce some potential caching-issue here in development mode:
so if you now change the title of a page and open a sibling of that
page, you won't see the updated title there until the sibling is
recompiled as well, which happens once you:
* restart the devMode
* hit save in the sibling-file
* edit any layout

i think blazing fast updates are well worth that downside.