remove RENDER_PATH_PATTERN
assuming that nobody will edit mesosphere-docs with a livereload-setup
anymore, we can now get rid of some complexity.

nobody needs to care about the rendering-pattern as by now the whole
DKP-docs compile within less than 5 seconds and re-compiles are almost
instant.

so i see no need to bug folks with that in the readme anymore.