don't bundle (most) mesosphere-pages in dev-image
we don't expect a lot of changes to mesosphere-content. let's ignore
that content when building the dev-container. this way the image gets
WAY! smaller and can be distributed without much of a hassle. it then
may serve as the single tool to use for folks writing docs.

note that we needed to strip a lot of (insignificant) references to the
mesosphere. these have been copy-pasted but have no impact on the
outcome.