apply handlebars-templating on all files
some files are repaired by this. e.g.:
https://docs.d2iq.com/mesosphere/dcos/2.2/installing/

currently has mustache-snippets in it that are not converted.

then there are some files that would render differently. those are the
ones changed in this commit.

THEN there are yet other files that contain `{{ }}`-snippets which will
not throw a warning. we'll "repair" those later, once we got a full
picture of how this templating should work. i think it's OK to have
authors escape those.