fix layout-compilation
those nifi-command-pages were broken (as in: could not be found on the
docs-page). there seems to be some maximum size for frontmatter-lines.

those errors popped up, now that we force everything through the
layouts-process! WOOHOO!

also we have a stray tutorial-sample-apps-page for dcos. no idea from
where it is linked to or who uses it, but i chose to leave it where it
is for now.