trim down the hierarchy plugin
we now build the tree of pages WAY more efficiently. additionally we add
a cache for pages that have the `menus`-meta (that we'll document
somewhen else).
also we got path-finding in a better shape (less guessing and regexing),
which gets rid of a weird plugin that we had inlined in index.js.

these measures speed up the layout-step from 40s to 6s on my machine for
a full build.