Private Beta
This commit is here to mark that the associated PR will be open for a while in order to provide a customer with our privte-beta-docs, as we don't want them to be fully public.

If we ever merge this branch back into `main` you'll want to remove THIS commit (or just revert its changes).

Here's a suggested workflow over the time we maintain this private beta:

### For each change:

1) First add the changes to the `main` branch! Not to this branch!
2) Merge/rebase `main` onto this branch to update the private beta docs.
  1) make sure that `git status` says "nothing to commit, working tree clean".
  2) run these git commands one by one and stop if something fails:
     ```shell-session
     git checkout main
     git pull origin main
     git checkout private-beta
     git pull origin private-beta
     git rebase main
     git push --force-with-lease
     ```
     If you run into trouble, please ask any developer for help.

     :warning: - if someone edits file on this branch and does not follow the workflow, there WILL be conflicts!

  3) Your changes are already being deployed. That might take up to 5 minutes.

### Once per product

1) on the main branch:
   Make sure the frontmatter of any product-version that shall be hidden from the general public has these attributes on the `main` branch:

   ```yaml
   subtree:
     beta: true
     draft: true
   ```

   * `subtree` tells the docs to put the sub-attributes to every child/descendant-page.
   * `beta` renders a big "beta"-badge right next to the title of a page.
   * `draft` tells the docs to not put this page onto the production-site.

   E.g. if we want to hide the docs for "Kommander 2.0", make sure that `/pages/dkp/kommander/2.0/index.md` has those attributes. This is the only file that needs those changes as it will hand down the needed attributes via `subtree`.

2) once a product-version has landed in `private-beta`:

   Make sure to remove the `draft: true` from the frontmatter of the main file.

   E.g. for "Kommander 2.0" you want to edit `/pages/dkp/kommander/2.0/index.md` like so:

   ```diff
   subtree:
     beta: true
   -  draft: true
   ```

   This will "unhide" this version and all child/descending-pages on this branch for customers to enjoy.

### When we're done with the private beta

In order to test whether we strictly followed the above process, please run a `git diff main private-beta`. The only changes you should see are a couple words in `README.md` and the abovementioned `draft: true`s that differ between the branches.

If you see more changes: work with the docs-team to find out what version needs to be on `main`.

If you don't see other changes, we strictly followed the above process, thus all changes that are in `private-beta` will also be in `main`. So we can now remove those `draft: true`s on the `main` branch and delete `private-beta` and the associated PR. Don't forget to tell customers to now go to the main docs-site or alternatively implement a redirect.