use develop branch for dev
we had a great deal of confusion this last release as folks would make
changes to main branch, hoping they would reflect on the production docs
site. I had to ask several people to close PRs or change the base
branch, and in some cases had to rewrite history to fix.

Lets use a new branch named develop that is more semantically in line
with the environment it maps to.

Also, removing scripts that are no longer in use as general cleanup