Edited readme, contributing, peg node version (#3941)
* Edited readme, contributing, peg node version

Making a lot of changes here to our docs to try and simplify it in
preparation for more contributors. Readme will be only for instructions
on how to use the repo and spin it up locally (and maybe some legacy
tasks like archiving)

Contributing file will be where we give guidelines for editing. Less
technical, more about how to edit and how to maintain the style guide.

Also, I've moved some of the git help into our private wiki, as this
won't be relevant to many of our contributors but only those unfamiliar
with git.

* change pull request template

* address vale and comments