fix: licensing policy and arn info (#4004)
* fix: licensing policy and arn info

Dimitri found when testing that I had forgotten to add the policy
information for consuming a license. Also, it was difficult to follow
where to get the license ARN from AWS.

Adding a prerequisite to the add licensing page because this is where
users are directed from the download page. The purchase page I felt was
superfluous and not linked anywhere, so I melded that into the existing
index.

Also, added a bit more color about how to find the license ARN. We want
to make sure we can test this end to end.

Closes D2IQ-82950
Closes D2IQ-82951

* docs: batch edits

Co-authored-by: Dimitri Koshkin <dimitri.koshkin@gmail.com>

Co-authored-by: Colin Colwell <85189139+Colin1841@users.noreply.github.com>
Co-authored-by: Dimitri Koshkin <dimitri.koshkin@gmail.com>
(cherry picked from commit 7873e0032dfee71c756fc6ad6c0867ba4be828f3)