Update index.md (#4420)
* Update index.md

All DKP commands on this page assume that "KUBECONFIG=clusterKubeconfig.conf" or that this correct context has been set in the default kubeconfig file.  We cannot assume this.  We need to either place a call-out at the top of the page to set the kubeconfig or place the proper "--kubeconfig=clusterkubeconfig.conf" entry in all the instructions.

* docs: markdown linting and adding note to subs

* Update index.md

* adding same note to air-gapped documentation

Co-authored-by: msdolbey <dolbes@gmail.com>
Co-authored-by: gvasquezvargas <98893586+gvasquezvargas@users.noreply.github.com>
Co-authored-by: Gabriela Vasquez <gvasquez@d2iq.com>